+++ opentelekomcloud/compute_deh_v1_hosts.go
@@ -1,0 +1,235 @@
+// This set of code handles all functions required to configure networking
+// on an opentelekomcloud_compute_instance_v2 resource.
+//
+// This is a complicated task because it's not possible to obtain all
+// information in a single API call. In fact, it even traverses multiple
+// OpenTelekomCloud services.
+//
+// The end result, from the user's point of view, is a structured set of
+// understandable network information within the instance resource.
+package opentelekomcloud
+
+import (
+	"fmt"
+	"github.com/gophercloud/gophercloud/openstack/compute/v2/servers"
+	"github.com/hashicorp/terraform/helper/schema"
+	"log"
+)
+
+// InstanceNICS is a structured representation of a Gophercloud servers.Server
+// virtual NIC.
+type InstanceNICS struct {
+	FixedIPv4 string
+	FixedIPv6 string
+}
+
+// InstanceAddresses is a collection of InstanceNICs, grouped by the
+// network name. An instance/server could have multiple NICs on the same
+// network.
+type InstancesAddress struct {
+	NetworkName  string
+	InstanceNICs []InstanceNICS
+}
+
+// InstanceNetwork represents a collection of network information that a
+// Terraform instance needs to satisfy all network information requirements.
+type InstanceNetworks struct {
+	UUID          string
+	Name          string
+	Port          string
+	FixedIP       string
+	AccessNetwork bool
+}
+
+// getAllInstanceNetworks loops through the networks defined in the Terraform
+// configuration and structures that information into something standard that
+// can be consumed by both OpenTelekomCloud and Terraform.
+//
+// This would be simple, except we have ensure both the network name and
+// network ID have been determined. This isn't just for the convenience of a
+// user specifying a human-readable network name, but the network information
+// returned by an OpenTelekomCloud instance only has the network name set! So if a
+// user specified a network ID, there's no way to correlate it to the instance
+// unless we know both the name and ID.
+//
+// Not only that, but we have to account for two OpenTelekomCloud network services
+// running: nova-network (legacy) and Neutron (current).
+//
+// In addition, if a port was specified, not all of the port information
+// will be displayed, such as multiple fixed and floating IPs. This resource
+// isn't currently configured for that type of flexibility. It's better to
+// reference the actual port resource itself.
+//
+// So, let's begin the journey.
+func getAllInstanceNetwork(d *schema.ResourceData, meta interface{}) ([]InstanceNetworks, error) {
+	var instanceNetworks []InstanceNetworks
+
+	networks := d.Get("addresses").([]interface{})
+	for _, v := range networks {
+		network := v.(map[string]interface{})
+		networkID := network["uuid"].(string)
+		networkName := network["name"].(string)
+		portID := network["port"].(string)
+
+		if networkID == "" && networkName == "" && portID == "" {
+			return nil, fmt.Errorf(
+				"At least one of network.uuid, network.name, or network.port must be set.")
+		}
+
+		// If a user specified both an ID and name, that makes things easy
+		// since both name and ID are already satisfied. No need to query
+		// further.
+		if networkID != "" && networkName != "" {
+			v := InstanceNetworks{
+				UUID:          networkID,
+				Name:          networkName,
+				Port:          portID,
+				FixedIP:       network["fixed_ip_v4"].(string),
+				AccessNetwork: network["access_network"].(bool),
+			}
+			instanceNetworks = append(instanceNetworks, v)
+			continue
+		}
+
+		// But if at least one of name or ID was missing, we have to query
+		// for that other piece.
+		//
+		// Priority is given to a port since a network ID or name usually isn't
+		// specified when using a port.
+		//
+		// Next priority is given to the network ID since it's guaranteed to be
+		// an exact match.
+		queryType := "name"
+		queryTerm := networkName
+		if networkID != "" {
+			queryType = "id"
+			queryTerm = networkID
+		}
+		if portID != "" {
+			queryType = "port"
+			queryTerm = portID
+		}
+
+		networkInfo, err := getInstanceNetworkInfo(d, meta, queryType, queryTerm)
+		if err != nil {
+			return nil, err
+		}
+
+		v := InstanceNetworks{
+			UUID:          networkInfo["uuid"].(string),
+			Name:          networkInfo["name"].(string),
+			Port:          portID,
+			FixedIP:       network["fixed_ip_v4"].(string),
+			AccessNetwork: network["access_network"].(bool),
+		}
+
+		instanceNetworks = append(instanceNetworks, v)
+	}
+
+	log.Printf("[DEBUG] getAllInstanceNetworks: %#v", instanceNetworks)
+	return instanceNetworks, nil
+}
+
+// getInstanceAddresses parses a Gophercloud server.Server's Address field into
+// a structured InstanceAddresses struct.
+func getInstancesAddress(addresses map[string]interface{}) []InstancesAddress {
+	var allInstanceAddresses []InstancesAddress
+
+	for networkName, v := range addresses {
+		instanceAddresses := InstancesAddress{
+			NetworkName: networkName,
+		}
+
+		instanceNIC := InstanceNICS{}
+		for _, v := range v.([]interface{}) {
+
+			v := v.(map[string]interface{})
+
+			if v["OS-EXT-IPS:type"] == "fixed" {
+				switch v["version"].(float64) {
+				case 6:
+					instanceNIC.FixedIPv6 = fmt.Sprintf("[%s]", v["addr"].(string))
+				default:
+					instanceNIC.FixedIPv4 = v["addr"].(string)
+				}
+			}
+
+			instanceAddresses.InstanceNICs = append(instanceAddresses.InstanceNICs, instanceNIC)
+		}
+
+		allInstanceAddresses = append(allInstanceAddresses, instanceAddresses)
+
+	}
+
+	log.Printf("[DEBUG] Addresses: %#v", addresses)
+	log.Printf("[DEBUG] allInstanceAddresses: %#v", allInstanceAddresses)
+
+	return allInstanceAddresses
+}
+
+// flattenInstanceNetworks collects instance network information from different
+// sources and aggregates it all together into a map array.
+func flattenInstanceNetwork(d *schema.ResourceData, meta interface{}) ([]map[string]interface{}, error) {
+
+	config := meta.(*Config)
+	computeClient, err := config.computeV2Client(GetRegion(d, config))
+	if err != nil {
+		return nil, fmt.Errorf("Error creating OpenTelekomCloud compute client: %s", err)
+	}
+
+	server, err := servers.Get(computeClient, d.Id()).Extract()
+	if err != nil {
+		return nil, CheckDeleted(d, err, "server")
+	}
+
+	allInstanceAddresses := getInstancesAddress(server.Addresses)
+	allInstanceNetworks, err := getAllInstanceNetwork(d, meta)
+	if err != nil {
+		return nil, err
+	}
+
+	networks := []map[string]interface{}{}
+
+	// If there were no instance networks returned, this means that there
+	// was not a network specified in the Terraform configuration. When this
+	// happens, the instance will be launched on a "default" network, if one
+	// is available. If there isn't, the instance will fail to launch, so
+	// this is a safe assumption at this point.
+	if len(allInstanceNetworks) == 0 {
+		for _, instanceAddresses := range allInstanceAddresses {
+			for _, instanceNIC := range instanceAddresses.InstanceNICs {
+				v := map[string]interface{}{
+					"name":        instanceAddresses.NetworkName,
+					"fixed_ip_v4": instanceNIC.FixedIPv4,
+				}
+				networks = append(networks, v)
+			}
+		}
+
+		log.Printf("[DEBUG] flattenInstanceNetworks: %#v", networks)
+		return networks, nil
+	}
+
+	// Loop through all networks and addresses, merge relevant address details.
+	for _, instanceNetwork := range allInstanceNetworks {
+		for _, instanceAddresses := range allInstanceAddresses {
+			if instanceNetwork.Name == instanceAddresses.NetworkName {
+				// Only use one NIC since it's possible the user defined another NIC
+				// on this same network in another Terraform network block.
+				instanceNIC := instanceAddresses.InstanceNICs[0]
+				copy(instanceAddresses.InstanceNICs, instanceAddresses.InstanceNICs[1:])
+				v := map[string]interface{}{
+					"name":           instanceAddresses.NetworkName,
+					"fixed_ip_v4":    instanceNIC.FixedIPv4,
+					"uuid":           instanceNetwork.UUID,
+					"port":           instanceNetwork.Port,
+					"access_network": instanceNetwork.AccessNetwork,
+				}
+				networks = append(networks, v)
+			}
+		}
+	}
+
+	log.Printf("[DEBUG] flattenInstanceNetworks: %#v", networks)
+	return networks, nil
+}