+++ opentelekomcloud/config.go
@@ -495,3 +495,10 @@ func (c *Config) autoscalingV1Client(region string) (*golangsdk.ServiceClient, e
 		Availability: c.getHwEndpointType(),
 	})
 }
+
+func (c *Config) dehV1Client(region string) (*golangsdk.ServiceClient, error) {
+	return huaweisdk.NewDeHServiceV1(c.HwClient, golangsdk.EndpointOpts{
+		Region:       c.determineRegion(region),
+		Availability: c.getHwEndpointType(),
+	})
+}